Added functionality to automatically skip pairing in case a bluetooth connection exists
When waiting for pairing we also check to see if a connection is already
made. If this is the case we simply stop trying to pair and continue
waiting for a bluetooth connection.

Once a connection is made we retreive the wifi data.
The sender of data can then decide when to try to connect over wifi.
Once that is completed we will return the wifi status. If it failed the
sender should send new data and try again